add package, documentation, test
