+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.util.TaskList;
+import duke.util.Ui;
+import duke.util.Storage;
+
 /**
  * The Command is the response of the program to be executed.
  * Execution of command may involve manipulation of TaskList,
