+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,13 @@
+package duke.command;
+
+import duke.util.TaskList;
+import duke.util.Ui;
+import duke.util.Storage;
+
+/**
+ * The exit command flags the program to exit upon execution.
+ * This terminates the chat bot.
+ */
 public class ExitCommand implements Command {
 
     public boolean isExit() {
