+++ src/main/java/duke/command/ListCommand.java
@@ -1,3 +1,13 @@
+package duke.command;
+
+import duke.util.TaskList;
+import duke.util.Ui;
+import duke.util.Storage;
+
+/**
+ * The list command helps to print all the tasks in the task list
+ * in a neatly-formatted way.
+ */
 public class ListCommand implements Command {
 
     @Override
