+++ src/main/java/duke/test/ParserTest.java
@@ -1,0 +1,19 @@
+package duke.test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import duke.util.Parser;
+
+class ParserTest {
+
+    @org.junit.jupiter.api.Test
+    void getTaskDescription() {
+        assertEquals("2 3 4", Parser.getTaskDescription(
+                new String[] {"1", "2", "3", "4"}));
+    }
+
+    @org.junit.jupiter.api.Test
+    void parse() {
+        // err ...
+    }
+}
